+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - Tests typo fix
+- Readme example
 
 ## [3.0.0] - 2019-09-12
 ### Changed