+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Logs in write methods (insert, multiInsert, update, save, multiSave, remove and multiRemove)
+
 ## [3.1.0] - 2019-10-09
 ### Fixed
 - Tests typo fix